Add LICENSE to .csproj and new release workflow
Added a reference to the LICENSE file in the `Piri.Extensions.DependencyInjection.csproj` file to include it in the package.

Introduced a new GitHub Actions workflow in `release.yml` to automate the release process to GitHub and potentially to NuGet. The workflow triggers on pushes to the `release` branch, checks out the code, sets up .NET, retrieves the latest tag, calculates a new version by incrementing the patch number, updates the version in the `.csproj` file, commits the change, and pushes the new version tag. It then packs the NuGet package and creates a new GitHub release with the new version. The step to publish to NuGet.org is included but commented out.